requirements.txt
Django==4.2.3
django-ckeditor==6.7.0
django-imagekit==4.1.0
django-jazzmin==2.6.0
django-js-asset==2.1.0
django-mptt==0.14.0
django-multiselectfield==0.1.12
pilkit==2.0
Pillow==10.0.0
django-polymorphic==3.1.0
